requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sqlalchemy
alembic
passlib
python-jose[cryptography]
bcrypt
python-dotenv
psycopg2-binary
uuid
email-validator
pydantic>=2.0
pydantic-settings
qrcode
google-auth 
google-auth-oauthlib 
google-auth-httplib2